feat: Separate logic and view for account password
